scripts/requirements.txt
torch
torchvision
numpy
scipy
Pillow
tensorboardX
tqdm
matplotlib
open3d
opencv-python